fix(guard): fold the re-export tests into test__symbols.py (PS-204 §2)
CI's audit-all reported a real error against MY file:

  [E] [PS-204 §2 orphan-test-file] tests/.../_guard/test__trials_reexport.py:
  no matching src file; mirror dir src/scitex_agent_container/_guard contains
  _check.py, _render.py, _report.py, _symbols.py, _trees.py — none match
  expected `_trials_reexport.py`

The rule wants an exact stem mirror, and there is no `_trials_reexport.py`
to mirror — the six tests assert that scripts/local_model_trials/detectors.py
re-exports the SAME OBJECTS the package defines, and all four of those
primitives live in _symbols.py. So _symbols.py is their mirror, and they
move into test__symbols.py unchanged.

Reproduced and verified locally: PS-204 is gone from audit-all, and the
guard suite is still 75 passed.

This is the half of the red that was mine. The rest of that job's failure
— audit-cli's [§4b] "help is free-form text — construct via CliHelp" and the
already-masked PS-226 job-name finding — names no file of this branch, is
shared with the other open PRs, and is deliberately NOT touched here: a
competing local fix or a suppression would silence a gate that is telling
the truth.
